bot auto-joins rooms when invited; new room-related commands
Any user can now invite the bot to a room of their choice. The
bot will then attempt to join the room automatically. Note that
encrypted rooms have not been tested yet and probably will not
work.

Furthermore, two room-related commands have been added: one to
show the rooms that the bot has joined, and one to force the bot
to leave a room. Both commands are only available to those users
that are allowed to stop the bot. That prevents "random" room
leaves and avoids to disclose room ids to any user.