fix bug in string splitting, add test cases for split() function
If the last element of a character-separated line is empty, it
will now be preserved in the result. For example, splitting the
string "a,b," at commas will now return { "a", "b", "" } instead
of the wrong { "a", "b" } vector.

Furthermore, some edge cases do not need to be checked, because
they are not really edge cases. Thus, the code is simpler now.