parse next_batch of sync requests in Sync::parse(), too
It should probably have been this way from the start. If there is
a dedicated method to parse the response of the Matrix sync
request (and there is), then that method should do all the
parsing. Furthermore, that also allows testing for the value of
next_batch. I see this as an absolute win.