fix possible type mixup in incidence calculation
The size type may be long unsigned int, but it may also be a
long long unsigned int. However, std::min() is templated and will
not find matching types when the size type is long long unsigned
int, so this has to be explicit and not assume long unsigned int.