introduce configuration option for synchronization delay
Up to version 0.4.1, the bot only checked for new Matrix events
every five seconds (plus overhead time for roundtrip to the
Matrix homeserver and the message / JSON parsing stuff). Now it
is possible to adjust that value to make the bot reply to the
commands faster or slower - within certain limits. Currently,
the minimum is 100 milliseconds, and the maximum is 30 seconds.
Configured values above or below that range will be changed to
be within that range.